Fullscreen all apps - no more subwindow

+++ internal_filesystem/main.py
@@ -48,9 +48,9 @@ drawer_open=False
 
 
 rootscreen = lv.screen_active()
-rootscreen.set_style_bg_color(lv.color_hex(0x444444), 0)
+#rootscreen.set_style_bg_color(lv.color_hex(0x444444), 0)
 rootlabel = lv.label(rootscreen)
-rootlabel.set_text("Welcome to PiggyOS!")
+rootlabel.set_text("Welcome!")
 rootlabel.align(lv.ALIGN.CENTER, 0, 0)
 
 def open_drawer():
@@ -295,17 +295,17 @@ def execute_script(script_source, is_file, is_launcher, is_graphical):
                 prevscreen = lv.screen_active()
                 newscreen=lv.obj()
                 newscreen.set_size(lv.pct(100),lv.pct(100))
-            timer1, timer2, timer3, timer4 = add_notification_bar(newscreen)
-            subwindow = lv.obj(newscreen)
-            subwindow.set_size(TFT_HOR_RES, TFT_VER_RES - NOTIFICATION_BAR_HEIGHT)
-            subwindow.set_pos(0, NOTIFICATION_BAR_HEIGHT)
-            subwindow.set_style_border_width(0, 0)
-            subwindow.set_style_pad_all(0, 0)
+            #timer1, timer2, timer3, timer4 = add_notification_bar(newscreen)
+            #subwindow = lv.obj(newscreen)
+            #subwindow.set_size(TFT_HOR_RES, TFT_VER_RES - NOTIFICATION_BAR_HEIGHT)
+            #subwindow.set_pos(0, NOTIFICATION_BAR_HEIGHT)
+            #subwindow.set_style_border_width(0, 0)
+            #subwindow.set_style_pad_all(0, 0)
             lv.screen_load(newscreen)
             script_globals = {
                 'lv': lv,
                 'appscreen': newscreen,
-                'subwindow': subwindow,
+                #'subwindow': subwindow,
                 'start_app': start_app, # for launcher apps
                 'parse_manifest': parse_manifest, # for launcher apps
                 'restart_launcher': restart_launcher, # for appstore apps
@@ -322,7 +322,7 @@ def execute_script(script_source, is_file, is_launcher, is_graphical):
             tb = getattr(e, '__traceback__', None)
             traceback.print_exception(type(e), e, tb)
         print(f"Thread {thread_id}: script {compile_name} finished")
-        if is_graphical and prevscreen and not is_launcher:
+        if False and is_graphical and prevscreen and not is_launcher: # disabled this for now
             print("/main.py: execute_script(): deleting timers...")
             timer1.delete()
             timer2.delete()
